Update local cypress run to match CI
Update the local portion of the cypress runner script so that the
ordering of starting the docker services is consistent. Adding the check
for exit code on some services should also help control the start up so
that all dependent services are running before the backend starts. This
should addressed an intermittent error where the backend would error on
initial connection to the db because it was attempting to the connect to
the db before the db migrations had completed.