[EASI-2902, EASI-2904, EASI-2916] - TRB text updates (#2001)
* Update link

* Snapshot

* Homepage text updates

* Remove reference to intake request form

* Empty commit - trigger PR workflow

* Removed unnecessary snapshot

The snapshot in this test was flaky due to the loading spinner within the request form.

We're already testing that the alert text exists, so the snapshot isn't necessary. The request form already has a snapshot in another test file.

* Update link href

If new request, go back to trb main tab. If request was already created, go back to task list.

* Fix missing ID in snapshot test

* Fix capitalization

Co-authored-by: Clay Benson <clay.benson@oddball.io>

---------

Co-authored-by: Clay Benson <clay.benson@oddball.io>